+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Anticipated release: July 9, 2021
 - Update the delete process for FFY ([#2996])
 - New design updates to Manage Account page ([#3204])
 - updated Assurance and Compliance text on APD review and export screen to show more detailed information. ([#3169]))
+- Login screen is disabled if api check fails ([#3223])
 
 #### 🐛 Bugs fixed
 
@@ -28,6 +29,7 @@ See our [release history](https://github.com/CMSgov/eAPD/releases)
 [#2581]: https://github.com/CMSgov/eAPD/issues/2581
 [#2996]: https://github.com/CMSgov/eAPD/issues/2996
 [#3030]: https://github.com/CMSgov/eAPD/issues/3030
+[#3223]: https://github.com/CMSgov/eAPD/issues/3223
 [#3181]: https://github.com/CMSgov/eAPD/issues/3181
 [#3204]: https://github.com/CMSgov/eAPD/issues/3204
 [#3207]: https://github.com/CMSgov/eAPD/issues/3207